Time: 14 ms (92.72%), Space: 42.1 MB (99.87%) - LeetHub

+++ 0115-distinct-subsequences/0115-distinct-subsequences.java
@@ -1,9 +1,9 @@
 class Solution {
     public int numDistinct(String s, String t) {
         int m=s.length(),n=t.length();
-        long[] dp = new long[n + 1];
+        int[] dp = new int[n + 1];
         
-        dp[0] = 1; // empty t
+        dp[0] = 1; 
         
         for (int i = 1; i <= m; i++) {
             for (int j = n; j >= 1; j--) {
@@ -12,6 +12,6 @@ class Solution {
                 }
             }
         }
-        return (int) dp[n];
+        return dp[n];
     }
 }
